Add website support to the application
Enhanced the MyApp component to include a new type for websites, commensurate adjustments have been made in styles and also imgUrl for appropriate display of the new type. Also, a MyWebsites component and a Websites view have been added to handle the loading and display of the websites in the application. This will support better categorization and accessibility of websites in the application. FilesMyApps.vue and MyApp.vue received majority of changes along with addition of new files and minor tweaks in other files.